irqchip: xilinx: Add support to remove Xilinx INTC irqchip driver module
commit b47db6fb92c7961f9d4f1a41d874b4feff56a148 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The existing irqchip implementation does not fully support use cases
where an irqchip driver has to be used as a module. In particular there
is no support to remove an irqchip driver module.
The use cases where an irqchip driver has to be loaded and then removed
as a module are really relevant in fpga world. A user can decide to
have a irqchip as part of a removable partial fpga region. In such cases
not only the corresponding irqchip driver has to be loaded as a module,
but must also be removed when the removable partial region is removed.

In the proposed implementation irqchip framework is updated to add a new
function platform_irqchip_remove through irqchip.h MODULE_DEVICE_TABLE
implementation.

The platform_irqchip_remove calls the corresponding interrupt controller
driver's remove routine. It is the responsibility of the interrupt
controller remove routine to ensure that all irq descriptors attached to
its irqdomain are freed up.
Changes in this patch is made through a new irqchip kconfig option with the
name IRQCHIP_XILINX_INTC_MODULE_SUPPORT_EXPERIMENTAL which is disabled
by default.

Signed-off-by: Anirudha Sarangi <anirudha.sarangi@xilinx.com>
Signed-off-by: Mubin Usman Sayyed <mubin.usman.sayyed@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>