net: ethernet: Clean Up Intel Axxia FEMAC driver
commit  b0162883a002934caf2231f181474523a0c7358e from
https://github.com/axxia/linux-yocto.git v5.10/standard/preempt-rt/axxia-dev/base

Remove support for PowerPC (3xxx) parts in this version of the driver.

Use dev_* instead of pr_* when possible.

Clean up stats and logging.

There were a log of messages like "...No end of packet...". In
most cases the packets were okay, but there were no buffers available.

Packets are also discarded, although by the driver and not the hardware,
when the destination MAC is not for us. This is the case because the
hardware only supports promiscuous mode.

The above will be logged using *_ratelimited or *_once functions.

Overflow, CRC, and alignment errors will be logged as they occur.

Change the default size of RX descriptor ring and buffers.

Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>