arm64: Workaround for Cortex-A53 erratum 1530924
commit f3837e88bc09190c13dc5f6e814d7ce617182b84 from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

Cortex-A53 erratum 1530924 allows TLB entries to be allocated as a
result of a speculative AT instruction. This may happen in the middle of
a guest world switch while the relevant VMSA configuration is in an
inconsistent state, leading to erroneous content being allocated into
TLBs.

The same workaround as is used for Cortex-A76 erratum 1165522
(WORKAROUND_SPECULATIVE_AT_VHE) can be used here.

Issue: ALB-7347
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>