mailbox: llce: Use 16 channels for logger
commit e86d630d707ebba1af423686e3cea67e19671a9e from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

This commit prepares the ground for LLCE logger driver which
will use SocketCAN interfaces to send the frames to user space.
For this purpose, the logger will use a channel for each interface
and NAPI specific commands (pop, release, enable / disable notifications)
to control the received data. It inherits all operations already
performed by LLCE CAN on RX except the local storage of a frame.

A local buffer for each interface was needed because all LLCE Logger
interfaces are receiving the frames over one single interrupt line and
the ID of the interface can be only identified after the frame is popped.

Issue: ALB-7505
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>