phy: xilinx: hdmi: Add linerate, nchannels and reset_gt
commit eb5a21a264d9528267cfb5e8815c88f6c43d1b13 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

For configuring HDMI 2.1, the line rate and the number of channels are
needed. These parameters are passed along with the hdmi21_conf
parameter.

The reset_gt is needed parameter is passed when there HDMI 2.1 Rx
subsystem driver has to reset the phy layer.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
State: pending
[michals: Squashed with
phy: Fix comments for nchannels reset_gt and linerate options
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>