phy: xilinx: hdmi: Add linerate, nchannels and reset_gt
commit c2081aaf1efc61090913125dd690a654747c94fc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

For configuring HDMI 2.1, the line rate and the number of channels are
needed. These parameters are passed along with the hdmi21_conf
parameter.

The reset_gt is needed parameter is passed when there HDMI 2.1 Rx
subsystem driver has to reset the phy layer.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
State: pending
[michals: Squashed with
phy: Fix comments for nchannels reset_gt and linerate options
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>