s32g274a: ddr: Fixed read_lpddr4_mr function
commit f4944282d5774c2320f21f41b23c79ce28e2b962 from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

The workaround for this errata should read only RANK 0
and check both dies.

Currently, the Linux driver only cleared rank0 bit, and reads MR from
RANK 1 which is not an issue on dual rank devices such as S32G274A-EVB,
but it is not correct on single rank devices, which do not have RANK 1.

Issue: ALB-7692
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>