tmu: s32gen1: Fixed bug about Kelvin to Celsius conversion and overflow.
commit ec3760c257e9cf9937e15965b8bfb1cbd1b5f2d5 from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

The value for TMU on s32gen1 was read incompletely as an 8-bit value
instead of a 9-bit one. Together with the fact that for s32gen1, unlike
for s32v234, the value is given in Kelvin and not in Celsius, it resulted
in errors of 17 Celsius degrees. This fixes it by fully reading the value
and converting it from Kelvin to Celsius degrees.

Issue: ALB-8014
Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>