phy: nxp-c45-tja11xx: add extts and perout support
commit 9da08e0c45214a808a8af2543fd03b467196a594 from
https://source.codeaurora.org/external/autobsps32/linux release/bsp32.0-5.10.90-rt

Add support for external timestamp and periodic signal output.
TJA1103 have one periodic signal and one external time stamp signal that
can be multiplexed on all 11 gpio pins.

The periodic signal can be only enabled or disabled. Have no start time
and if is enabled will be generated with a period of one second in sync
with the LTC seconds counter. The phase change is possible only with a
half of a second.

The external timestamp signal has no interrupt and no valid bit and
that's why the timestamps are handled by polling in .do_aux_work.

Issue: ALB-8254

Signed-off-by: Radu Pirea (NXP OSS) <radu-nicolae.pirea@oss.nxp.com>
Acked-by: Richard Cochran <richardcochran@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>