uio: Add the UIO driver for AI engine
commit a221710ea6a7410a7b07061a9a15bb84d5513b7e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

This adds the UIO driver for AI engine. The driver is
more or less a simple wrapper that enables the UIO dmem genirq
with some required configurations, because the UIO dmem driver
is not probed through DT compatible string. This driver initializes
the UIO dmem by adding it as a child device.

While doing it, any resources in this driver DT node is mirroed to
the child node so that those resources can be accessed by UIO
if needed. The child device is set up with parent node DMA
configuration, thus the child device can work with DMA mapping APIs.

In addition, the UIO dmem driver expects some information to be
specified in pdata given by the initializer, which includes,
     - uioinfo: version, name, irq type, mem offsets,,,
     - memory allocation information: size and number of chunks

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
State: pending
[Wendy: Replace old IRQ simulation APIs with the new ones in v5.10]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>