spi: spi-zynqmp-gqspi: Fix erase failure when tx buswidth is 4
commit c4032f80858cb8e4297ee4af3f3889e4a4e23262 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Irrespective of the tx buswidth, for erase commands the flash device
expects to receive the address on single line. But when tx buswidth
is set to 4. The driver while issuing erase commands, send the address
on 4 lines. This causes the erase operation to fail. This patch updates
the driver and sends the address, for all erase commands, on single line.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>