usb: dwc3: Added remote wake-up in xilinx glue driver
commit db6b55a74c709a27fdf6d175f890568dc8945424 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

To support USB remote wake-up in xilinx glue driver, Added disable
regulator interface in core to enter in U3(suspend) state, and
introduced wake-up capability flag to enable the feature. when the
remote wakeup flag is enabled and the controller is in U3 state
(suspended), it requires suspend_clk to be active to detect wake-up
signaling driven on the link.

Added function registration and deregistration functionality for
remote wakeup feature from/to dwc3-host.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>