v4l: xilinx: dprx: Add MMCM configuration for rx_dec_clk for versal systems
commit fa6b52d6751f090c04e1441415abfbbac8ad8480 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

A Clocking Wizard is added which has an MMCM in the DPRX subsystem to
provide the rx_dec_clk. This is required to give the link_rate/20 clock
for 8B10B logic which is included in DPRXSS for versal systems. This
patch adds supported code to configure the MMCM based on the link rate.

Signed-off-by: Jagadeesh Banisetti <jagadeesh.banisetti@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>