misc: xilinx-ai-engine: device struct for tile type
commit 72d673074fed7c544f1365bfb2ac448710230c5f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

add aie device structure as an argument to the get tile type
operation. For future aie devices, there can be a variation in the
number of memory tile rows available in the device. Adding this as a
property to the aie device will allow the driver to scale for different
combinations of tile types.

Signed-off-by: Tejus Siddagangaiah <tejus.siddagangaiah@xilinx.com>
Acked-by: Pramod Peethambaran <pramod.peethambaran@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>