v4l: xilinx: axis-subsetconv: Add driver for axis subset converter
commit b919e94ad5b07632bdc21ecb069e297e9b650211 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

This converter driver is for matching the format of source pad and
sink pad in the media pipeline. The format of a source pad does
not match the sink pad if it is converted by a non-memory mapped
hardware IP. This subset converter driver is for non-memory mapped
axi stream subset converter which converts the format of the stream.

Refer to Xilinx PG085 document for IP information.

Signed-off-by: Anil Kumar Mamidala <anil.mamidala@xilinx.com>
Reviewed-by: Vishal Sagar <vishal.sagar@xilinx.com>
State: pending
[michals:
v6.6
xilinx: Convert MODULE_LICENSE GPL v2 to GPL
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>