staging: uartlite-rs485: Initial commit to add drv
commit 1d7b60e56a55543882cc88d7f19a1bfb13fe9890 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Add staging driver which can support rs485 over axi uartlite.
This is duplicate of drivers/tty/serial/uartlite.c, with added
support for rs485 when the IP is patched to drive the DE signal.

The duplicated code is modified to avoid any conflict with the
original driver and fix warning by renaming/changing following:
- device name
- driver name
- Major number
- compatibility string
- Remove all the console related code as it will never be used.
- Minor space/spellcheck fixes to remove the checkpatch warnings.

Consider this as temporary stop gap solution and should be removed from
staging area as soon as original IP and driver supports this feature.

Signed-off-by: Pallav Joshi <pallav.joshi@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>