edac: xilinx: make event manager registration configurable
commit db7b2f38879fb589489636a64f96644433767008 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Currently the event manager does not support multiple registrations.
The default is event manager is enabled. To move to interrupts in the
driver xilinx-ddrmc-edac.disable_event=1 is to be passed.

We can pass module parameters as arguments in the builtin driver.
bootargs = "console=ttyAMA0  earlycon=pl011,mmio32,0xFF000000,115200n8
	    clk_ignore_unused init_fatal_sh=1 xilinx-ddrmc-edac.disable_event=1";

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>