crypto: xilinx: Add upstreamed SHA driver
commit 6029c6209f24de64feaeaf1b8bab98402fb722f2 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This patch adds the upstreamed ZU+ SHA driver as zynqmp-sha.c. The current
driver is renamed as zynqmp-sha-deprecated.c and it will be maintained for
two releases and then deprecated. By default the old driver will be built.
An option is added in Kconfig to select the new driver if required. The
Kconfig and Makefile is also modified to reflect this update.

Signed-off-by: Harsha <harsha.harsha@xilinx.com>
State: upstream (7ecc3e34474b, 7a70d9a1cf11)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
[Quanyang: delete modifications in config files.]