mtd: spi-nor: For erase size equal to 0 set erase size_shift to 0
commit 42160d51674db96a881ff8145ef14423cbdf3793 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

While computing the subset of supported 4Byte opcodes for sector erase
commands, for non supported erase commands the opcode and erase size are
set to 0xFF & 0x00 respectively. With erase size set to 0x00
erase->size_sfit is computed as 0xFFFFFFFF. When UBSAN is enabled in
kernel config this results in shift-out-of-bounds error and kernel call
trace. Fix the issue by setting erase->size_shift  to 0 when erase size
is 0.

Addresses-UBSAN: shift-out-of-bounds

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>