mtd: spi-nor: swp: fix unlock lower area failure
commit 338b5d9ebcbcb45e77ca20c63e1adc3d9f2132af from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

When we try to unlock a locked region in the lower area of flash, the
variables "can_be_top" and "can_be_bottom" may be all true and this is
not right because for this case "lock_len" is only "ofs" but not
"mtd->size - (ofs + len)".

So introduce spi_nor_is_lower_area() to check if the area to be unlocked
is in the lower and use it to exclude "can_be_top".

It's similar to spi_nor_is_upper_area() which is used to exclude
"can_be_bottom".

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>