fpga: zynqmp: Make word align the configuration data
commit 07ea2247e93b8f8b37710722bc33506a932562e8 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

To avoid the unwanted copies at firmware(PMUFW) this patch provides
a fix to align PL configuration data if the data is not word-aligned.
It improves the PL configuration timings to process unaligned bit
files.

To align the configuration data this patch adds a few padding bytes and
these additional padding bytes will not create any functional impact on
the PL configuration.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>