soc: xilinx: fix unhandled SGI warning message
commit 0f13387206a018a11472688e7342b4e9796bfc68 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Xen broadcasts SGI to each VM when multiple VMs run on Xen hypervisor. In
such case spurious SGI is expected if one event is registered by one VM and
not registered by another VM. We let users know that Unhandled SGI is not
error and expected if kernel is running on Xen hypervisor.

Signed-off-by: Tanmay Shah <tanmay.shah@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>