usb: add quirk for xhci reinit warning on resume
commit bef009a2e22479af1c4f2feffd85b81abc5d9a81 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Add quirk 'snps,xhci-reset-on-resume' to resolve the warning generated
on resume, during the suspend cycle when USB is not pluh-in or USB not
set as a wakeup source then power management turn off the clock and
power for usb source, to save power.

Added quirk will address suspend-resume use-case, and reset the host
controller on resume.

Fixed for warning:
xhci-hcd xhci-hcd.2.auto: xHC error in resume, USBSTS 0x401, Reinit

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>