usb: add quirk for xhci reinit warning on resume
commit 48fd8f458c25479b02d96a5fda6c567af52685dd from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Add quirk 'snps,xhci-reset-on-resume' to resolve the warning generated
on resume, during the suspend cycle when USB is not pluh-in or USB not
set as a wakeup source then power management turn off the clock and
power for usb source, to save power.

Added quirk will address suspend-resume use-case, and reset the host
controller on resume.

Fixed for warning:
xhci-hcd xhci-hcd.2.auto: xHC error in resume, USBSTS 0x401, Reinit

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>