arm: dts: zynqmp: add snps,xhci-reset-on-resume quirk
commit 0e85ae97b4cd7290bd73239a21eb2de1a819465a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Add quirk 'snps,xhci-reset-on-resume' to resolve the warning generated
on resume, during the suspend cycle when USB is not pluh-in or USB not
set as a wakeup source then power management turn off the clock and
power for usb source, to save power.

Quirk added for both usb0 and usb1.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>