i3c: master: dw: HACK: Update the scl timings
commit d760717fa4caec372ec167580b163c121e003775 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Update the SCL Push Pull and Open Drain timing configuration.
Hard coded SCL configuration values are used as temporary hack.
Push Pull high and low count  is set to 5 Open Drain high and low count set to 32

The patch is workaround for slow systems but it will be resolved in near
future.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>