usb: dwc3: xilinx: add reset-controller support for keystoneb
commit 2606bd766d8ef7f622ebedb33fbef4f647feaafb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Add a reset-controller for supporting Xilinx keystoneb platforms.
Trigger the reset, with the assert and deassert reset controller
APIs for USB controller initialization. Delay of microseconds is
added in between assert and deassert to meet the setup and hold
the time requirement of the reset.

Modify the code to put the keystoneb usb controller into either
D3 or D0 state with reset controller APIs.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>