Revert "arm64: zynqmp: Comment sata phy setting for zcu102"
commit 80a615e164ed14ad6130f93e3c9d73db7c190a22 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This reverts commit 728139491742c76576e2c02227fd8f21714e2665.

'phys' DT property, containing a phandle to the phy port node and a device
type for device configuration. When this property is not defined in DT,
then suspend-resume use-case failed. Without failure, after the resume,
SATA drive should be detected in the system.

Reverting for the SATA support for zcu102 board to address the above
suspend-resume use case.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>