i2c: xiic: Fix Rx and Tx paths in standard mode
commit a58db5c64f0e57ddb07c9577656a272286dea318 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

When a combined message request comes from user space, the
controller has to initiate repeated start sequence. In standard
mode, this repeated start sequence is corrupted if there is still
data in the Tx FIFO.

So, always make sure that all the bytes are completely transmitted
out of the FIFO by waiting for TXEMPTY, if the previous message is
of Tx type.

Signed-off-by: Raviteja Narayanam <raviteja.narayanam@xilinx.com>
Signed-off-by: Manikanta Guntupalli <manikanta.guntupalli@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Link: https://lore.kernel.org/all/1663836294-5698-3-git-send-email-manikanta.guntupalli@xilinx.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>