i2c: xiic: Switch to Xiic standard mode for i2c-read
commit 60e5e0190ef373128a4cf9db9c4206b19769ea56 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Xilinx I2C IP has two modes of operation, both of which implement
I2C transactions. The only difference from sw perspective is the
programming sequence for these modes.
Dynamic mode  -> Simple to program, less number of steps in sequence.
Standard mode -> Gives flexibility, more number of steps in sequence.

In dynamic mode, during the i2c-read transactions, if there is a
delay(> 200us) between the register writes (address & byte count),
read transaction fails. On a system with load, this scenario is
occurring frequently.
To avoid this, switch to standard mode if there is a read request.

Added a quirk to identify the IP version effected by this and follow
the standard mode.

Signed-off-by: Raviteja Narayanam <raviteja.narayanam@xilinx.com>
Signed-off-by: Manikanta Guntupalli <manikanta.guntupalli@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Link: https://lore.kernel.org/all/1663836294-5698-4-git-send-email-manikanta.guntupalli@xilinx.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>