spi: spi-cadence-quadspi: Add support for odd byte count read/write in DDR mode
commit ce3dad5e683b2f164b8aa8b72918f7d5167b7845 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Due to the nature of DDR protocol, an even number of bytes is always
transferred and also the least significant bit of the byte address
shall always be zero when using DDR protocol.
This patch add support for read and write operations for the
requests with odd byte count or odd address or both. These kind of requests
are expected with UBIFS file system and mtd_stresstest.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>