spi: spi-cadence-quadspi: Fix ospi resume failures
commit e119d3abac0066ccf0b6f4ea27deed0dd40ce8e5 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

MTD suspend handler in spi-nor disables the Octal DDR mode and enables
it again during the resume path, handling of this sequence is not proper
inside ospi driver (spi-cadence-quadspi).
This patch takes care of switching the mode from Octal DDR to SDR and
vice-versa, also by the time ospi resume handler called, OSPI flash
device is in Octal DDR mode and controller is in SDR mode. This creates
an undefined behaviour as controller and device are in different modes, so
update the ospi resume handler to take care of switching back both OSPI
flash device and controller to SDR mode.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>