net: xilinx: Fix phylink_of_phy_connect error on RGMII designs
commit 9d0ee11ba44b842abbd1a1e40b935c69defa3ce5 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

For PCS/PMA based designs where PHY mode can be sgmii or 1000BaseX,
the existing implementation checks for pcs-handle and phy-handle
properties. For other modes (e.g RGMII), the implementation does
not check for these properties which results in PHY failures.

Update the code to ensure that pcs-handle and phy-handle properties
are parsed for all cases except XXV and MRMAC which do not have PHY
in the designs.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>