Revert "fpga: region: Adds runtime PM support"
commit 0b77df34eb95c6c66c9cbdf5355d61f410101d0a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

The existing solution is added based on the internal request and it's
capable of handling only single power-domain and it may not cover some
use cases that of-fpga-region driver supports. fpga-region driver is
capable of handling multiple fpga-regions and it may need multiple
power domains. Also, it is possible there are already a devices in
fpga-region for static OF tree, or an overlay with no 'firmware-name'
but 'external-fpga-config.' In this case, the existing solution failed
to handle the relevant power-domains.

Due to the above drawbacks with the existing solution revert the below
commit ID's "30ff73ae40e90263e8ba20a6640381add54dedf6" and
"7239a6a4314dcda4583bb8e090517f0be37736bf".
This change will not impact any of the existing FPGA use cases.

To add runtime PM handling support to the FPGA subsystem needs to get
possible suggestions to overcome the above drawbacks from the upstream.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>