firmware: xilinx: add bitmask support for get_op_char API in feature check
commit 8c8d8542eb59f3bba953f69238ae29b36de7f7c3 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

As per the current code base in the firmware, the power
characteristics are there in ZynqMP but not in Versal. Also, the
temperature characteristics are there in Versal but not in ZynqMP.

In order to avoid these conflicts added the bitmask feature in the
feature check API. Now, the user can query the supported
functionality first and can know whether it is supported or not in
the firmware. So accordingly the user can perform their actual
functitonality across the different platforms.

Signed-off-by: Ronak Jain <ronak.jain@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>