net: ethtool: netlink: introduce ethnl_update_bool()
commit 90539312a6763a3459db432dd28a51f19c8cc20c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Due to the fact that the kernel-side data structures have been carried
over from the ioctl-based ethtool, we are now in the situation where we
have an ethnl_update_bool32() function, but the plain function that
operates on a boolean value kept in an actual u8 netlink attribute
doesn't exist.

With new ethtool features that are exposed solely over netlink, the
kernel data structures will use the "bool" type, so we will need this
kind of helper. Introduce it now; it's needed for things like
verify-disabled for the MAC merge configuration.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Link: https://lore.kernel.org/all/20230119122705.73054-2-vladimir.oltean@nxp.com/
Signed-off-by: Pranavi Somisetty <pranavi.somisetty@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: upstream (7c494a7749a7)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>