net: ethtool: add helpers for MM fragment size translation
commit 0988998113704c976e8487840f890fc7e643a8aa from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

We deliberately make the Linux UAPI pass the minimum fragment size in
octets, even though IEEE 802.3 defines it as discrete values, and
addFragSize is just the multiplier. This is because there is nothing
impossible in operating with an in-between value for the fragment size
of non-final preempted fragments, and there may even appear hardware
which supports the in-between sizes.

For the hardware which just understands the addFragSize multiplier,
create two helpers which translate back and forth the values passed in
octets.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Link: https://lore.kernel.org/all/20230119122705.73054-8-vladimir.oltean@nxp.com/
Signed-off-by: Pranavi Somisetty <pranavi.somisetty@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: upstream (dd1c41645039)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>