mtd: spi-nor: Fix Flash protection failure on ISSI 1G flash parts
commit 1683f2b7b527bf479526252c203b987e4db20155 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

During flash lock/unlock the ISSI 1g QSPI flashes supports top/bottom
protection, but the top/bottom bit is present in the function register
& not in the status register. While performing the flash lock/unlock for
enabling/disabling the top/bottom the driver was instead updating the BP3
bit in the status register. Due to this unwanted memory blocks were getting
locked, which resulted in flash protection failures.
Update the flash protection flags of the flashes to avoid writing the
top/bottom information to the BP3 bit in the status register.

Fixes: ee5164a4135d ("mtd: spi-nor: Avoid locking/unlocking unrequested memory blocks")
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>