usb: dwc3: otg: remove default otg state B_IDLE configuration
commit 3e7637bc10f0a4ebf764d37db5e44625a76e485f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

During the OTG initialization, OTG driver sets OTG state OTG_STATE_IDLE as
part of gadget initialization, which can lead wrong configuration compare
to ID pin the state.

Added workaround, to address the incorrect configuration, start the OTG
state machine with the UNDEFINED state and initiate OTG functionality with
the correct state of the ID pin at the start of the main thread.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>