misc: xilinx-ai-engine: Add error interrupt support for aie2 hw
commit 8136a2276ef9077eb34374c0d5e1413baf68aa17 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Adds support for error interrupt handling for aie2 hw. aie2 and above
has mem tiles and it needs to be handled differently. Broadcast event
(bc_event) bit is different for mem tiles and aie tiles. For mem tiles we
should not check the bc_event from aie tile and break. So tile backtrack
needs to be separate for aie tiles and mem tiles.

Signed-off-by: Govindarajulu Varadarajan <govindarajulu.varadarajan@amd.com>
Acked-by: Gregory Williams <gregory.williams@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>