spi: spi-cadence: Fix transfer timeout issue
In spi-cadence during a transfer the driver, immediately after pushing all
the remaining data to TXFIFO, reads the RXFIFO data, if the RXFIFO doesn't
have the data when the controller tries to read it, then that results in a
spi timeout error. Update the driver to wait for the TXFIFO to empty before
going for RXFIFO read operation and a 10us is added before reading the data
from RXFIFO as we cannot use RXNEmpty interrupt for reading data from
RXFIFO as there is a known IP issue (AR:65885).

Fixes: 48dbbe728e74 ("spi: spi-cadence: Add support for Slave mode")
Signed-off-by: Srinivas Goud <srinivas.goud@amd.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending