drm: xlnx: hdmi: Remove unused hdmi->wait_for_streamup check in xlnx_hdcp_key_store()
commit 27e943fb0c01248cde299711c2b6a0a857a85d95 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

xlnx_hdcp_key_store() will return the number of keys it has configured
successfully. Otherwise, it returns an error if the HDMI stream is not
ready or in any other failure cases.

The flag wait_for_streamup check is useless as this variable is always
'0' in the driver. The HDMI driver is not updating this variable in any
conditions. So, this condition is removed because the HDCP code flow
performs these checks in respective api's already.

Signed-off-by: Lakshmi Prasanna Eachuri <lakshmi.prasanna.eachuri@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>