xilinx: firmware: Add sysfs to get last reset reason
commit dbcf61f655164dbf85f43bfbdd0da7f2ff09bc11 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add sysfs interface to get the last reset reason of the system from
a user space.

When the system reset happens whether intentionally or due to some
errors, the firmware used to store the last reset reason and the same
be queried by using the IOCTL API or the sysfs entry to know the
reason behind the system reset and accordingly, the error can be
diagnostic.

Signed-off-by: Ronak Jain <ronak.jain@amd.com>
Link: https://lore.kernel.org/r/20231121112056.1762641-3-ronak.jain@amd.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>