dt-bindings: usb: dwc3: Add snps,enable_guctl1_ipd_quirk
commit 87d70c18cf2ffb1b3a1e227db6c1407625649f43 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

SNPS controller when configureed in HOST mode maintains Inter Packet
Delay (IPD) of ~380ns which works with most of the super-speed hubs
except VIA-LAB hubs. When IPD is ~380ns HOST controller fails to
enumerate FS/LS devices when connected behind VIA-LAB hubs.

To address the above issue, added 'snps,enable_guctl1_ipd_quirk' quirk,
This quirk set the bit 9 of GUCTL1 that enables the workaround in HW to
reduce the ULPI clock latency by 1 cycle, thus reducing the IPD (~360ns).

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
[Piyush: In the zynqmp database bit 9 which is reserved but we used as a
chicken bit to enable a fix related to Inter Packet Delay in HW. This is
this is a ZU+ specific so it should be part of internal patches.
]
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>