firmware: xilinx: correct last reset reason api arguments
commit 921f961962d4c74083f7aa81b929882672964c79 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

The last reset reason API is not available in the upstream and during
6.6 kernel rebase the below change got merged to add support for
additional arguments for PM EEMI APIs. Hence due to this change the
arguments are mismatched for the last reset reason API and it is
failing.

Corrected the arguments properly to fix the issue.

Upstream commit f922b16aa5fa ("firmware: xilinx: Update firmware call
interface to support additional args") added support for additional args.

Fixes: 9be9fbac7853 ("firmware: xilinx: Add support for GET_LAST_RESET_REASON IOCTL")
Signed-off-by: Ronak Jain <ronak.jain@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>