arm64: zynqmp: Fix kr260 clock wiring
commit 001421dbd68a39416029d987b37b8c126237b840 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

kr260 revA/revA01 is using discrete oscilator for DP (27MHz) and si5332 for
other clocks but clocks are different compare to kv260 that's why fix it to
aligned with the latest schematics.

On the other handle kr260 revB/revA03 also contains 74.25 MHz discrete
clock chip for SLVC-EC output which is not defined.

Link: https://lore.kernel.org/r/e87ae94979c6efc909740bb1a569505042e4f876.1706626255.git.michal.simek@amd.com
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>