mtd: spi-nor: Store GPIO info for resetting device during resume
commit 482a084ce752a24644c696544a82f3b7f1ee201e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Upon resume, SPI-NOR triggers the hardware reset API to reset the flash
device by toggling the GPIO. This step is necessary because after resume,
the flash needs to be put in SDR mode. To facilitate this, SPI-NOR needs
GPIO information to reset the flash device. Therefore, a gpio_desc member
is added to the spi_nor structure to retain the GPIO information and
utilize it for resetting the device.

Fixes: 58713004fad0 ("spi: spi-cadence-quadspi: Add support for OSPI device reset")
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>