v4l: xilinx: dprx: Replace irq_of_parse_and_map with platform_get_irq_byname
commit 6abdfc41ef9ac57fb937ec602a283c787c158d1f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

HDCP protocols HDCP1X and HDCP2X are optional. The order of HDCP
interrupts will get changed based on enablement of HDCP1X and HDCP2X.

When HDCP2X is enabled, the index of HDCP1X will be 3, otherwise the
index of HDCP1X is '1'.

So, the driver should get IRQ for a device by irq-name instead of using it
with fixed index.

Signed-off-by: Lakshmi Prasanna Eachuri <lakshmi.prasanna.eachuri@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>