phy: xilinx-xhdmiphy: Fix MMCM parameter values for GTYE5/GTYP
commit 991c014e6b7ce6829ca1b3c0c155fabbf782c13f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Currently there is no clock being supplied to VTC in the HDMI 2.1 Tx
Subsystem. This is due to link training failure which occurs because
of unstable clock produced from MMCM block for GTYE5. The unstable
clock is a result of incorrect parameters being set in driver. Fix
this by setting the correct parameters for GTYE5

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>