spi: Raise limit on number of chip selects
commit e003a60c5d1501d007316d1641ed9565db4f8bd7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

As reported by Guenter the limit we've got on the number of chip selects is
set too low for some systems, raise the limit. We should really remove the
hard coded limit but this is needed as a fix so let's do the simple thing
and raise the limit for now.

Fixes: 4d8ff6b0991d ("spi: Add multi-cs memories support in SPI core")
Reported-by: Guenter Roeck <linux@roeck-us.net>
Suggested-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: https://msgid.link/r/20240124-spi-multi-cs-max-v2-1-df6fc5ab1abc@kernel.org
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>