Revert "usb: gadget: webcam: Add support for Full HD and 4K UHD Raw Video Formats"
commit 727905d1793e8ff3e66b77b4013541b8e946d80d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

This reverts commit 0cc61c993b9cb81a7e970bcae029bcfc2cefbe55.

The support was added for VCU TRD UVC usecase, but this feature is no
longer required and validated due to TRD end-of-life. So revert it and
be in sync with the mainline linux kernel.

If in future there is a valid usecase then this has to be first
discussed with upstream community and then added back.

Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>