remoteproc: zynqmp: Add coredump support
commit 9cc1d7cacb5c5e29b86bfb154afa9d1e639550a2 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Supporting remoteproc coredump requires the platform-specific driver to
register coredump segments to be dumped. Do this by calling
rproc_coredump_add_segment for every carveout.

Also call rproc_coredump_set_elf_info when then rproc is created. If the
ELFCLASS parameter is not provided then coredump fails with an error.
Other drivers seem to pass EM_NONE for the machine argument but for me
this shows a warning in gdb. Pass EM_ARM because this is an ARM R5.

Signed-off-by: Leonard Crestez <cdleonard@gmail.com>
Link: https://lore.kernel.org/r/d4556268-8274-4089-949f-3b97d67793c7@gmail.com
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
State: upstream (fec2601f2003)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>