firmware: xilinx: fix feature check logic for TF-A specific APIs
commit c071ec4e12b632463231581aa4d3f26331a93fe7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Currently, Linux is passing only API ID for feature of TF-A specific
APIs but for feature check of TF-A specific APIs, TF-A is checking for
Module ID + API ID as a result incorrect version is received for all TF-A
specific APIs. So, fix feature check logic to pass valid arguments to
get correct version.

Signed-off-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>