v4l: xilinx: isppipeline: Add support for AWB
commit 06b15f3a82c2ceb4b4bfb3df0c6500f07a749b84 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add support for Auto White Balance ISP configuration.

Auto White Balance scales the values of pixels based on a gray-world
assumption which states that the average of all channels should result
in a gray image. It adds a modification which thresholds pixels based
on their saturation value and only uses pixels below the provided
threshold in finding average pixel values.

A threshold of 1 means that all pixels are used to white-balance,
while a threshold of 0 means no pixels are used. Lower thresholds
are useful in white-balancing saturated images.

"awb_threshold" ranges from 0 to 1 is the configurable option.

Signed-off-by: Mounik Katikala <mounik.katikala@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>