v4l: xilinx: isppipeline: Add support for DEGAMMA
commit e3d5b663e810d6e0bb89e57ec0cd8b88562b8988 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add support for DEGAMMA ISP configuration.

Degamma also known as linearization is typically designed to linearize
the input from a sensor (or any pre-processing IP) in order to facilitate
ISP processing that operates on a linear domain, such as noise reduction.
While non-linearization comes in several forms, this module is served to
linearize only simple curves.

This implementation supports Bayer Raw data of 8 and 16 bits
per pixel with a maximum of 64 knee points PWL mapping.

"select_degamma" ranges from 0 to 1 is the configurable option.

Signed-off-by: Mounik Katikala <mounik.katikala@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>