v4l: xilinx: isppipeline: Add support for LSC
commit 9478a8914b59d0e9c6b8b1c35cf70bdb8a34c086 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add support for Lens Shading Correction ISP configuration.

Vignetting/Lens shading refers to the fall-off pixel intensity
from the centre towards the edges of the image. In this algorithm,
vignette is corrected by considering the distance between the centre
pixel and actual image pixel position. This distance is used to
calculate intensity gain per pixel per channel which is used for the
correction.

Signed-off-by: Mounik Katikala <mounik.katikala@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>