v4l: xilinx: isppipeline: Add support for RESIZE
commit 05567fe946ad254df4a309a296206d0e64e68857 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add support for RESIZE ISP configuration.

Resolution Conversion is the method used to resize the source image to
the size of the destination image. Different types of interpolation
techniques can be used in resize function, namely: Nearest-neighbor,
Bilinear, and Area interpolation. The type of interpolation can be
passed as a template parameter to the API. The following enumeration
types can be used to specify the interpolation type:

-  XF_INTERPOLATION_NN - For Nearest-neighbor interpolation
-  XF_INTERPOLATION_BILINEAR - For Bilinear interpolation
-  XF_INTERPOLATION_AREA - For Area interpolation

Note: Scaling factors greater than or equal to 0.25 are supported in
down-scaling and values less than or equal to 8 are supported for
up-scaling.

Signed-off-by: Mounik Katikala <mounik.katikala@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>