arm64: zynqmp: Fix pwm-fan polarity
commit 501c10e0cfdaea551299b6be09381d759ccab77d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

The correct operating mode for the fan is inversed (1). The
previous pwm driver implementation had a bug and the polarity
information was propagated incorrectly to the kernel. The normal (0)
polarity specified in the device tree was incorrectly clearing the
polarity bit in the counter control register. After the bug fix,
setting the polarity to inversed (1) in the device tree will clear
the polarity bit.

Signed-off-by: Vishal Patel <vishal.patel@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>