staging: xilinx-tsn: fix kernel crash when new bridge if is created
commit d4ebc0fc2f7c7eb3615ef9e769d4bcc36115ff21 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

A kernel crash was noticed during the creation of a new bridge
interface, occurring due to the handling of switchdev events. This is
due to the event handler receiving net_device data belonging to the
actual bridge interface. The driver was trying to interpret the
net_device private data as axienet_local and causing the crash.

To address this problem, two checks have been incorporated into the
driver. The first check ensures that the upper device acts as a master,
while the second one verifies if the net_device is managed by the
driver. This approach is a standard procedure for managing switchdev
events. If the device isn't managed by the driver, it's expected to
return NOTIFY_DONE, allowing the switchdev framework to continue with
other devices.

Signed-off-by: Nagadheeraj Rottela <nagadheeraj.rottela@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>