i2c: cadence: Add atomic transfer support for controller version 1.4
commit 4b81772652594341086843c53e44f2559c99ff14 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Rework the read and write code paths in the driver to support operation
in atomic contexts in master mode. This change does not apply to slave
mode because there is no way to handle interruptions in that context.

Adjust the message timeout to include some extra time. For non-atomic
contexts, 500 ms is added to the timeout. For atomic contexts,
2000 ms is added because transfers happen in polled mode, requiring
more time to account for the polling overhead.

Similar changes have been implemented in other drivers, including:
commit 3a5ee18d2a32 ("i2c: imx: implement master_xfer_atomic callback")
commit 445094c8a9fb ("i2c: exynos5: add support for atomic transfers")
commit ede2299f7101 ("i2c: tegra: Support atomic transfers")
commit fe402bd09049 ("i2c: meson: implement the master_xfer_atomic
callback")

Signed-off-by: Manikanta Guntupalli <manikanta.guntupalli@amd.com>
Link: https://lore.kernel.org/linux-arm-kernel/20240801094408.2004460-4-manikanta.guntupalli@amd.com/
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>