crypto: asymmetric_keys - Adjust signature size calculation for NIST P521
commit 1eb37d55a6b76d06e58f9aef5cca476e693fd1b1 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Adjust the calculation of the maximum signature size for support of
NIST P521. While existing curves may prepend a 0 byte to their coordinates
(to make the number positive), NIST P521 will not do this since only the
first bit in the most significant byte is used.

If the encoding of the x & y coordinates requires at least 128 bytes then
an additional byte is needed for the encoding of the length. Take this into
account when calculating the maximum signature size.

Reviewed-by: Lukas Wunner <lukas@wunner.de>
Tested-by: Lukas Wunner <lukas@wunner.de>
Reviewed-by: Jarkko Sakkinen <jarkko@kernel.org>
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
State: upstream (4dc50330a437e91063c68da074574bf6459d68a5)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>