staging: xilinx-tsn: Add Endpoint Packet Switch Feature and Control Frame Handling for STP/LLDP
commit af6254f7bb2ed9d87ce0f019d041cbaa92ef1179 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Introduce changes to support the Endpoint Packet Switch feature and
enchance the handling of Layer-2 point-to-point control frames.

Enabled management frame handling based on the presence of the
`xlnx,packet-switch` property in the device tree.
Configured the switch to recognize and handle management frames,
specifically those with the destination MAC address `01-80-C2-00-00-00`,
used by STP (802.1d) and LLDP. If `xlnx,packet-switch` is enabled, these
frames are forwarded exclusively to the Endpoint port. Otherwise,
frames are forwarded to all ports.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>