staging: xilinx-tsn: Enable Traffic Shaping Support with QDisc Framework
commit 6e8f9645251b460ca6f1c4e5e8d66c0a2b41a936 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

QDisc (Queueing Discipline) framework in the Linux kernel is used to
implement traffic shaping functionality into the Xilinx TSN
(Time-Sensitive Networking) driver. The axienet_tsn_shaper_tc function
is introduced, allowing for an easy setup of Traffic Control (TC) for
Traffic Shaping in the driver. Furthermore, hardware offload support
is activated for processing TC configurations efficiently.

The current implementation of the code is designed to support
two or three queues for the TSN IP.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Message-ID: <20240927093230.2126496-1-srinivas.neeli@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>