staging: xilinx-tsn: Add IRQ-safe spinlock to protect TX BD processing
commit 73e317e33cba79ce2eb1cdbc8f5ad3e74bcf1732 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

TADMA transmit and TADMA transmit post processing are already protected
with spinlocks. Any areas with shared critical sections, such as alm and
register access are covered under this. However there is one section in
the IRQ code that performs read operations over shared memory to decide
count for tx post processing. There are no modifications to critical
shared resources here,but it is safer to protect this under spinlock
when tadma_xmit() and irq handler threads are running in parallel.

Suggested-by: Timor Knudsen <timor.knudsen@amd.com>
Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Message-ID: <20240930054200.2741713-1-srinivas.neeli@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>