arm64: versal: Describe EEPROM on i2c on vek280
commit d46cc2f1e144ac2ecad43904fee3fd596948a3d6 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

i2c bus is shared between DUT and SC(System Controller) and it is not
recommended to use any device connected to i2c mux. In case of EEPROM which
is placed on i2c bus (without i2c mux) it shouldn't be a problem to
describe it. EEPROM is read only that's why it shouldn't be a problem to
access it via i2c multimaster bus from both sides.

Signed-off-by: Michal Simek <michal.simek@amd.com>
Message-ID: <6e7e238e53bdb1b2f5cdd4d455c5388960cbd9ed.1727686739.git.michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>