dt-bindings: net: xlnx,axi-ethernet: Add bindings documentation for Ethernet Offload Engine
commit 9c972453935eaeea36f66f6c26650b8e87acc1f0 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add devicetree bindings documentation for Ethernet Offload Engine.

The ethernet offload engine support features offloaded from software to
the hardware to increase the performance and efficiency of the system.
These offloaded features include calculating checksum and segmentation
offloading of the packet if its size is greater than MTU and in Rx, it
supports the Receive Offload feature.

It is an independent IP and can be integrated into existing ethernet
subsystem IPs.

Co-developed-by: Kiran kumar Gandham <kirankumar.gandham@amd.com>
Signed-off-by: Kiran kumar Gandham <kirankumar.gandham@amd.com>
Signed-off-by: Ravikanth Tuniki <ravikanth.tuniki@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Message-ID: <20241007072418.2623055-4-ravikanth.tuniki@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>