net: axienet: Add EOE GSO and checksum support for TX
commit 4aaf1b42ffe689b22bb6fba44d4d11f4c2f26d3a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Generic Segmentation Offload performs the segmentation of the
UDP packet i.e. creating multiple small packets of 1472 bytes
from a larger packet of up to 64KB size. Tx coalesce count is
set to 1. With offload enabled, there are not as many interrupts
as before and the interrupt for every 64KB segment needs to be
handled immediately to ensure better performance.

Signed-off-by: Kiran kumar Gandham <kirankumar.gandham@amd.com>
Signed-off-by: Ravikanth Tuniki <ravikanth.tuniki@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Message-ID: <20241007072418.2623055-6-ravikanth.tuniki@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>