dmaengine: zynqmp_dma: Fix the reset value of dma channels
commit d2b8c13d1bb7d7417600cb4de2b8498165f3788c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

DMA channel registers should be in default state before Linux driver is
probed. The correct way to ensure this is to issue a reset. However, this
is problematic because the physical reset is common to all 8 ZynqMP DMA
channels. The solution to this is to pass SW reset IDs from SW to
firmware. Firmware should check for channel owners (APU or RPU) and use
either the physical reset line or clear the IP registers to ensure that
the channels are in default state. This functionality is not supported
in the firmware. So as a workaround set the channel control register to
the default state.

Signed-off-by: Abin Joseph <abin.joseph@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Message-ID: <20241007172614.3039215-1-abin.joseph@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>