usb: misc: onboard_usb_dev: skip suspend/resume sequence for USB5744 SMBus support
commit 535bccda6ee9b45651929e9f959d03e8256262cf from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

USB5744 SMBus initialization is done once in probe() and doing it in resume
is not supported so avoid going into suspend and reset the HUB.

There is a sysfs property 'always_powered_in_suspend' to implement this
feature but since default state should be set to a working configuration
so override this property value.

It fixes the suspend/resume testcase on Kria KR260 Robotics Starter Kit.

Fixes: 75d7a6531b44 ("usb: misc: onboard_usb_dev: add Microchip usb5744 SMBus programming support")
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Message-ID: <1729238864-3362294-1-git-send-email-radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>